rename occurrences of volto-interaktiv-alttextgenerator to @interaktivgmbh/volto-alttextgenerator
