+++ packages/@interaktivgmbh/volto-alttextgenerator/.gitignore
