+++ packages/@interaktivgmbh/volto-alttextgenerator/.release-it.json
