+++ packages/@interaktivgmbh/volto-alttextgenerator/babel.config.js
