+++ packages/@interaktivgmbh/volto-alttextgenerator/locales/de/LC_MESSAGES/volto.po
