+++ packages/@interaktivgmbh/volto-alttextgenerator/locales/volto.pot
