+++ packages/@interaktivgmbh/volto-alttextgenerator/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@interaktivgmbh/volto-alttextgenerator",
+  "version": "1.0.0",
+  "description": "The Volto add-on for interaktiv.alttextgenerator",
+  "main": "src/index.ts",
+  "license": "MIT",
+  "keywords": [
+    "volto-addon",
+    "volto",
+    "plone",
+    "react",
+    "accessibility",
+    "ai"
+  ],
+  "author": "Interaktiv GmbH",
+  "homepage": "https://github.com/interaktivgmbh/volto-alttextgenerator#readme",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:interaktivgmbh/volto-alttextgenerator"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "i18n": "rm -rf build/messages && NODE_ENV=production i18n --addon",
+    "dry-release": "release-it --dry-run",
+    "release": "release-it",
+    "release-major-alpha": "release-it major --preRelease=alpha",
+    "release-alpha": "release-it --preRelease=alpha"
+  },
+  "addons": [],
+  "peerDependencies": {
+    "@plone/volto": "^16.34.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "@plone/scripts": "^3.0.0",
+    "@plone/types": "^1.5.0",
+    "@types/react": "^18.3.1",
+    "@types/react-dom": "^18.3.1",
+    "release-it": "^19.0.5",
+    "typescript": "^5.7.3"
+  }
+}
