+++ packages/@interaktivgmbh/volto-alttextgenerator/src/actions/alttexts/alttexts.ts
@@ -1,0 +1,13 @@
+import { PATCH_ALT_TEXT_SUGGESTION } from '@interaktivgmbh/volto-alttextgenerator/constants/ActionTypes';
+import type { AltTextSuggestionAction } from '@interaktivgmbh/volto-alttextgenerator/types';
+
+export function updateAltTextSuggestion(path: string): AltTextSuggestionAction {
+  return {
+    type: PATCH_ALT_TEXT_SUGGESTION,
+    request: {
+      op: 'patch',
+      path: `${path}/@alt_text_suggestion`,
+      data: { path },
+    },
+  };
+}
