+++ packages/@interaktivgmbh/volto-alttextgenerator/src/config/settings.ts
