+++ packages/@interaktivgmbh/volto-alttextgenerator/src/constants/ActionTypes.ts
