+++ packages/@interaktivgmbh/volto-alttextgenerator/src/customizations/volto/components/manage/Blocks/HeroImageLeft/Data.jsx
