+++ packages/@interaktivgmbh/volto-alttextgenerator/src/index.ts
