+++ packages/@interaktivgmbh/volto-alttextgenerator/src/types/index.ts
