--- packages/volto-interaktiv-alttextgenerator/src/messages.ts
@@ -1,44 +1,0 @@
-import { defineMessages } from 'react-intl';
-import type { Messages } from 'volto-interaktiv-alttextgenerator/types';
-
-const messages: Messages = defineMessages({
-  altTextGenStartLabel: {
-    id: 'An alternative text is being generated in the background.',
-    defaultMessage: 'An alternative text is being generated in the background.',
-  },
-  altTextGenSuccessTitle: {
-    id: 'Success',
-    defaultMessage: 'Success',
-  },
-  altTextGenSuccessLabel: {
-    id: 'The alternative text was generated successfully.',
-    defaultMessage: 'The alternative text was generated successfully.',
-  },
-  aiGeneratedAltLabel: {
-    id: 'Was this alt text generated by an AI?',
-    defaultMessage: 'Was this alt text generated by an AI?',
-  },
-  altTextGenErrorTitle: {
-    id: 'Error',
-    defaultMessage: 'Error',
-  },
-  altTextGenUnsupported: {
-    id: 'The generation of alt texts is not supported for this image type.',
-    defaultMessage:
-      'The generation of alt texts is not supported for this image type.',
-  },
-  altTextGenNotAllowed: {
-    id: 'The generation of alt texts is disabled for this content.',
-    defaultMessage: 'The generation of alt texts is disabled for this content.',
-  },
-  altTextGenErrorLabel: {
-    id: 'The alt text generation failed.',
-    defaultMessage: 'The alt text generation failed.',
-  },
-  altTextIsAIGenerated: {
-    id: 'Alternative text generated by AI',
-    defaultMessage: 'Alternative text generated by AI',
-  },
-});
-
-export default messages;
