--- packages/volto-interaktiv-alttextgenerator/towncrier.toml
@@ -1,33 +1,0 @@
-[tool.towncrier]
-filename = "CHANGELOG.md"
-directory = "news/"
-title_format = "## {version} ({project_date})"
-underlines = ["", "", ""]
-template = "./node_modules/@plone/scripts/templates/towncrier_template.jinja"
-start_string = "<!-- towncrier release notes start -->\n"
-issue_format = "[#{issue}](https://github.com/interaktivgmbh/volto-interaktiv-alttextgenerator/issue/{issue})"
-
-[[tool.towncrier.type]]
-directory = "breaking"
-name = "Breaking"
-showcontent = true
-
-[[tool.towncrier.type]]
-directory = "feature"
-name = "Feature"
-showcontent = true
-
-[[tool.towncrier.type]]
-directory = "bugfix"
-name = "Bugfix"
-showcontent = true
-
-[[tool.towncrier.type]]
-directory = "internal"
-name = "Internal"
-showcontent = true
-
-[[tool.towncrier.type]]
-directory = "documentation"
-name = "Documentation"
-showcontent = true
