--- packages/volto-interaktiv-alttextgenerator/tsconfig.json
@@ -1,30 +1,0 @@
-{
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "target": "es2022",
-    "allowJs": true,
-    "resolveJsonModule": true,
-    "moduleDetection": "force",
-    "isolatedModules": true,
-    "verbatimModuleSyntax": true,
-    "module": "preserve",
-    "noEmit": true,
-    "lib": ["es2022", "dom", "dom.iterable"],
-    "jsx": "react-jsx",
-    "paths": {
-      "@plone/volto/*": ["../../core/packages/volto/src/*"],
-      "volto-interaktiv-alttextgenerator/*": ["./src/*"]
-    }
-  },
-  "include": ["**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx"],
-  "exclude": [
-    "node_modules",
-    "build",
-    "public",
-    "coverage",
-    "**/*.test.{js,jsx,ts,tsx}",
-    "**/*.spec.{js,jsx,ts,tsx}",
-    "**/*.stories.{js,jsx,ts,tsx}"
-  ]
-}
